+++ apps/rds_mysql/outputs.tf
@@ -15,5 +15,6 @@ output "db_admin_username" {
 }
 
 output "db_admin_password" {
+  sensitive = true
   value = "${aws_db_instance.database.password}"
 }